FIX: make if-statement for extra-informative message in evaluate_connect_function more permissive
the if-statement in this function is very specific and ensures that the error message both starts with "global name" and ends with "is not defined" before raising the informative error about nested imports. https://github.com/nipreps/nibabies/pull/365 gives one example where this if-statement is too specific, and doesn't catch a NameError that does actually arise from a module-level import.